rollup/x/hm/types/message_submit_answer_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"hm/testutil/sample"
)

func TestMsgSubmitAnswer_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgSubmitAnswer
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgSubmitAnswer{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgSubmitAnswer{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
